[heapdump] Use usize instead of u64 for program addresses
By definition, usize is big enough to contain every address on the
target platform.

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1710190
Commit-Queue: Fabio D'Urso <fdurso@google.com>

+++ src/performance/memory/heapdump/instrumentation/src/hooks.rs
@@ -23,7 +23,7 @@ const STACK_TRACE_MAXIMUM_COMPRESSED_SIZE: usize =
 static QUICK_EARLY_RETURN: AtomicBool = AtomicBool::new(false);
 
 unsafe extern "C" {
-    fn __sanitizer_fast_backtrace(buffer: *mut u64, buffer_size: usize) -> usize;
+    fn __sanitizer_fast_backtrace(buffer: *mut usize, buffer_size: usize) -> usize;
 }
 
 pub fn enable_quick_early_return() {
